error fix: upon searching with decimals

+++ app/src/main/java/com/kakeibo/feature_main/domain/use_cases/use_case_search/InsertSearchUseCase.kt
@@ -45,7 +45,7 @@ class InsertSearchUseCase(
                 )
             }
 
-            if (searchModel.fromAmount!!.toLong() > searchModel.toAmount!!.toLong()) {
+            if (searchModel.fromAmount!!.toBigDecimal() > searchModel.toAmount!!.toBigDecimal()) {
                 throw SearchEntity.InvalidSearchException(
                     context.getString(R.string.err_min_amount_greater)
                 )
